:zap: erc1155: Change ticker_nonce type to be included in JS's max safe integer

+++ erc1155/definition/src/state.rs
@@ -78,7 +78,7 @@ pub struct State {
     pub settings: Settings,
 
     pub default_token: String,
-    pub ticker_nonce: u64,
+    pub ticker_nonce: u32,
     pub tokens: HashMap<String, Token>,
     pub approvals: HashMap<String, Approvals>,
